// TheSharks.UI/src/types/application/activity/Participant.ts
export interface Participant {
    registratorFirstName: string;
    registratorLastName: string;
    registree: string;
    registreeId: string;
    registreeEmail: string;
    registreePhoneNumber: string;
    diveCertificate: string;
    remark: string;
    id: string;
    asDiver: boolean;
    openWaterTestTitle: string;
    openWaterTestContent: string;
}