// TheSharks.UI/src/types/application/role/Claims.ts
export enum Claims {
    ManageMembers = "Beheer Leden",
    ManageActivities = "Beheer Activiteiten",
    ManageNewsItems = "Beheer Nieuwsitems",
    ManageDownloadables = "Beheer Downloads",
    ManageGalleries = "Beheer Gallerijen",
    ManagePageContent = "Beheer Inhoud Website",
    ManageStatistics = "Statistieken"
}